refactor: move git page widget into page feature
